requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.32
langchain==0.2.13
langchain-core==0.2.31
langchain-community==0.2.12
pytest==8.3.2
httpx==0.27.0

python-multipart==0.0.9
textblob==0.17.1
